+++ app/Http/Controllers/Admin/EnglishWordCrudController.php
@@ -51,15 +51,7 @@ class EnglishWordCrudController extends CrudController
     {
         CRUD::setFromDb(); // columns
 
-        $this->crud->addField([
-            'label'             => 'Categories',
-            'type'              => 'select2_multiple',
-            'name'              => 'categories', // the method that defines the relationship in your Model
-            'entity'            => 'categories', // the method that defines the relationship in your Model
-            'attribute'         => 'name', // foreign key attribute that is shown to user
-            'model'             => WordCategory::class,
-            'pivot'             => true, // on create&update, do you need to add/delete pivot table entries?
-        ]);
+        CRUD::removeColumn('incorrect_answers');
         /**
          * Columns can be defined using the fluent syntax or array syntax:
          * - CRUD::column('price')->type('number');
@@ -77,48 +69,80 @@ class EnglishWordCrudController extends CrudController
     {
         CRUD::setValidation(EnglishWordRequest::class);
 
-        CRUD::setFromDb(); // fields
-
-        CRUD::removeField('incorrect_answers');
-
-        $this->crud->addField([
-            'label'             => 'Categories',
-            'type'              => 'select2_multiple',
-            'name'              => 'categories', // the method that defines the relationship in your Model
-            'entity'            => 'categories', // the method that defines the relationship in your Model
-            'attribute'         => 'name', // foreign key attribute that is shown to user
-            'model'             => WordCategory::class,
-            'pivot'             => true, // on create&update, do you need to add/delete pivot table entries?
-        ]);
-
-        $this->crud->addField([
-            'name' => 'incorrect_answers',
-            'Label' => 'Ҷавобҳои нодурст!',
-            'type' => 'repeatable',
-            'fields' => [
-                [
-                    'name' => 'incorrect_english_1',
-                    'label' => 'Чавоби нодуруст',
-                    'type' => 'text',
-                    'wrapper' => [
-                        'class' => 'form-group col-md-6'
-                    ],
-                ],
-                [
-                    'name' => 'incorrect_english_2',
-                    'label' => 'Чавоби нодуруст',
-                    'type' => 'text',
-                    'wrapper' => [
-                        'class' => 'form-group col-md-6'
-                    ],
-                ]
+        $this->crud->addFields([
+            [
+                'name' => 'english',
+                'type' => 'text',
+                'label' => 'Луғати Англиси',
             ],
-            // optional
-            'min_rows' => 1, // minimum rows allowed, when reached the "delete" buttons will be hidden
-            'max_rows' => 1, // maximum rows allowed, when reached the "new item" button will be hidden
+            [
+                'name' => 'tj',
+                'type' => 'text',
+                'label' => 'Луғати тоҷики',
+            ],
+            [
+                'label'             => 'Категория луғатҳо',
+                'type'              => 'select2_multiple',
+                'name'              => 'categories', // the method that defines the relationship in your Model
+                'entity'            => 'categories', // the method that defines the relationship in your Model
+                'attribute'         => 'name', // foreign key attribute that is shown to user
+                'model'             => WordCategory::class,
+                'options' => (function($query) {
+                    return $query->whereNotIn('slug', WordCategory::SPECIALS_SLUGS)->get();
+                }),
+                'pivot'             => true, // on create&update, do you need to add/delete pivot table entries?
+            ],
+            [
+                'name' => 'incorrect_answers',
+                'label' => 'Ҷавобҳои нодурст!',
+                'type' => 'repeatable',
+                'fields' => [
+                    [
+                        'name' => 'incorrect_word_1',
+                        'label' => 'Чавоби нодуруст',
+                        'type' => 'text',
+                        'wrapper' => [
+                            'class' => 'form-group col-md-6'
+                        ],
+                    ],
+                    [
+                        'name' => 'incorrect_word_2',
+                        'label' => 'Чавоби нодуруст',
+                        'type' => 'text',
+                        'wrapper' => [
+                            'class' => 'form-group col-md-6'
+                        ],
+                    ]
+                ],
+                // optional
+                'min_rows' => 1, // minimum rows allowed, when reached the "delete" buttons will be hidden
+                'max_rows' => 1,
+            ],
+            [
+                'name' => 'is_noun',
+                'type' => 'checkbox',
+                'label' => 'Исм',
+                'wrapper' => [
+                    'class' => 'form-group col-md-2'
+                ],
+            ],
+            [
+                'name' => 'is_verb',
+                'type' => 'checkbox',
+                'label' => 'Феъл',
+                'wrapper' => [
+                    'class' => 'form-group col-md-2'
+                ],
+            ],
+            [
+                'name' => 'is_adjective',
+                'type' => 'checkbox',
+                'label' => 'Сифат',
+                'wrapper' => [
+                    'class' => 'form-group col-md-2'
+                ],
+            ],
         ]);
-
-
         /**
          * Fields can be defined using the fluent syntax or array syntax:
          * - CRUD::field('price')->type('number');
@@ -145,10 +169,13 @@ class EnglishWordCrudController extends CrudController
 
 
         $dto = new EnglishWordDto(
-            $request->english,
+            ucfirst($request->english),
             $request->tj,
             $request->categories,
-            $request->incorrect_answers
+            $request->incorrect_answers,
+            $request->is_noun,
+            $request->is_verb,
+            $request->is_adjective,
         );
 
         $word = $this->englishWordService->store($dto);
@@ -174,10 +201,13 @@ class EnglishWordCrudController extends CrudController
         $request = $this->crud->getRequest();
 
         $dto = new EnglishWordDto(
-            $request->english,
+            ucfirst($request->english),
             $request->tj,
             $request->categories,
-            $request->incorrect_answers
+            $request->incorrect_answers,
+            $request->is_noun,
+            $request->is_verb,
+            $request->is_adjective,
         );
 
         $word = $this->englishWordService->update($dto, $request->id);