+++ app/Http/Requests/RussianWordRequest.php
@@ -20,8 +20,16 @@ class RussianWordRequest extends FormRequest
 
     protected function prepareForValidation()
     {
+        $incorrect_answers = json_decode($this->incorrect_answers, true);
+        foreach ($incorrect_answers[0] as $key => $value){
+            if(trim($value) == '') {
+                $incorrect_answers = [];
+                break;
+            }
+        }
         $this->merge([
-            'incorrect_answers' => json_decode($this->incorrect_answers, true)
+            'incorrect_answers' => $incorrect_answers,
+            'russian' => trim($this->russian)
         ]);
     }
 
@@ -34,10 +42,16 @@ class RussianWordRequest extends FormRequest
     public function rules()
     {
         return [
-            'russian' => 'required',
+            'russian' => 'required|unique:russian_words,russian,'.$this->id,
             'tj' => 'required',
             'categories' => 'required',
             'incorrect_answers' => 'required',
+            'is_masculine' => 'boolean',
+            'is_feminine' => 'boolean',
+            'is_neutral' => 'boolean',
+            'is_noun' => 'boolean',
+            'is_verb' => 'boolean',
+            'is_adjective' => 'boolean'
         ];
     }
 
@@ -61,10 +75,11 @@ class RussianWordRequest extends FormRequest
     public function messages()
     {
         return [
-            'word.required' => 'Луғатро нависед',
-            'translate.required' => 'Тарҷумаро нависед',
+            'russian.required' => 'Луғатро нависед',
+            'russian.unique' => 'Луғат алакай сохта шудааст',
+            'tj.required' => 'Тарҷумаро нависед',
             'categories.required' => 'Категория луғатро интхоб кунед',
-            'words.required' => 'Луғатҳои нодурустро интихоб кунед',
+            'incorrect_answers.required' => 'Луғатҳои нодурустро интихоб кунед',
         ];
     }
 }