+++ app/Services/RussianWordService.php
@@ -18,7 +18,12 @@ class RussianWordService {
         $word->russian = $request->getRussian();
         $word->tj = $request->getTj();
         $word->incorrect_answers = $request->getIncorrectAnswers();
-
+        $word->is_masculine = $request->isMasculine();
+        $word->is_feminine = $request->isFeminine();
+        $word->is_neutral = $request->isNeutral();
+        $word->is_noun = $request->isNoun();
+        $word->is_verb = $request->isVerb();
+        $word->is_adjective = $request->isAdjective();
 
         if($word->save()){
             $word->categories()->sync($request->getCategories());
@@ -42,6 +47,12 @@ class RussianWordService {
         $word->russian = $request->getRussian();
         $word->tj = $request->getTj();
         $word->incorrect_answers = $request->getIncorrectAnswers();
+        $word->is_masculine = $request->isMasculine();
+        $word->is_feminine = $request->isFeminine();
+        $word->is_neutral = $request->isNeutral();
+        $word->is_noun = $request->isNoun();
+        $word->is_verb = $request->isVerb();
+        $word->is_adjective = $request->isAdjective();
 
         if($word->update()){